refactor: migrating variants-overview snippets
