+++ src/components/DataExplorerContent.js
@@ -1,6 +1,5 @@
 import React, {useEffect} from "react";
-import {useSelector} from "react-redux";
-
+import { useDispatch, useSelector } from "react-redux";
 import {Redirect, Route, Switch} from "react-router-dom";
 
 import ExplorerGenomeBrowserContent from "./explorer/ExplorerGenomeBrowserContent";
@@ -10,10 +9,16 @@ import ExplorerSearchContent from "./explorer/ExplorerSearchContent";
 import {SITE_NAME} from "../constants";
 import {withBasePath} from "../utils/url";
 
+import { performGetGohanVariantsOverviewIfPossible } from "../modules/explorer/actions";
+
 
 const DataExplorerContent = () => {
+    const dispatch = useDispatch();
     useEffect(() => {
         document.title = `${SITE_NAME} - Explore Your Data`;
+
+        // Trigger Gohan variants-overview fetch
+        dispatch(performGetGohanVariantsOverviewIfPossible());
     }, []);
 
     const chordURL = useSelector(state => state.nodeInfo.data?.CHORD_URL);
