+++ src/modules/discovery/actions.js
@@ -2,7 +2,6 @@ import {createNetworkActionTypes, networkAction} from "../../utils/actions";
 
 
 export const PERFORM_GOHAN_GENE_SEARCH = createNetworkActionTypes("GOHAN_GENE_SEARCH");
-export const PERFORM_GET_GOHAN_VARIANTS_OVERVIEW = createNetworkActionTypes("GET_GOHAN_VARIANTS_OVERVIEW");
 
 
 
@@ -18,18 +17,4 @@ const performGohanGeneSearch = networkAction((searchUrl) => () => ({
     types: PERFORM_GOHAN_GENE_SEARCH,
     url: searchUrl,
     err: "error performing Gohan gene search",
-}));
-
-
-export const performGetGohanVariantsOverviewIfPossible = () => (dispatch, getState) => {
-    const gohanUrl = getState()?.services?.gohan?.url;
-    const bentoBaseUrl = `${getState().nodeInfo.data.CHORD_URL}`;
-    const overviewPath = `/variants/overview`;
-    const getUrl = gohanUrl ? `${gohanUrl}${overviewPath}` : `${bentoBaseUrl}api/gohan${overviewPath}`;
-    dispatch(performGetGohanVariantsOverview(getUrl));
-};
-const performGetGohanVariantsOverview = networkAction((getUrl) => () => ({
-    types: PERFORM_GET_GOHAN_VARIANTS_OVERVIEW,
-    url: getUrl,
-    err: "error getting Gohan variants overview",
 }));
