+++ src/modules/discovery/reducers.js
@@ -1,12 +1,10 @@
 import {
     PERFORM_GOHAN_GENE_SEARCH,
-    PERFORM_GET_GOHAN_VARIANTS_OVERVIEW
 } from "./actions";
 
 export const discovery = (
     state = {
         geneNameSearchResponse: [],
-        variantsOverviewResponse: {},
     },
     action
 ) => {
@@ -16,12 +14,7 @@ export const discovery = (
                 ...state,
                 geneNameSearchResponse: action.data.results
             };
-
-            case PERFORM_GET_GOHAN_VARIANTS_OVERVIEW.RECEIVE:
-            return {
-                ...state,
-                variantsOverviewResponse: action.data
-            };            
+          
 
         default:
             return state;
