+++ src/modules/explorer/reducers.js
@@ -12,6 +12,7 @@ import { readFromLocalStorage } from "../../utils/localStorageUtils";
 import { DEFAULT_OTHER_THRESHOLD_PERCENTAGE } from "../../constants";
 
 import {
+    PERFORM_GET_GOHAN_VARIANTS_OVERVIEW,
     PERFORM_SEARCH,
     PERFORM_INDIVIDUAL_CSV_DOWNLOAD,
     ADD_DATA_TYPE_QUERY_FORM,
@@ -29,6 +30,7 @@ import {
 // TODO: Could this somehow be combined with discovery?
 export const explorer = (
     state = {
+        variantsOverviewResponse: {},
         dataTypeFormsByDatasetID: {},
         fetchingSearchByDatasetID: {},
         searchResultsByDatasetID: {},
@@ -47,6 +49,11 @@ export const explorer = (
     action
 ) => {
     switch (action.type) {
+        case PERFORM_GET_GOHAN_VARIANTS_OVERVIEW.RECEIVE:
+            return {
+                ...state,
+                variantsOverviewResponse: action.data
+            };  
         case PERFORM_SEARCH.REQUEST:
             return {
                 ...state,
